+++ tests/test_workflow/tests/test_parent_workflow.py
@@ -27,7 +27,6 @@ def test_hello_world_parent_workflow(workflow_data, workflow_runner):
     }
     workflow_runner(
         "parent_workflow.wdl",
-        "hello_world_workflow",
         inputs,
         expected
     )
